requirements.txt
Pillow>=10.0.0  # For image processing
requests>=2.31.0  # For fetching CSS files
beautifulsoup4>=4.12.0  # For parsing HTML/CSS
colorthief>=0.2.1  # For extracting dominant colors
webcolors>=1.13.0  # For color name mapping
numpy>=1.24.0  # For color calculations
scikit-learn>=1.4.0  # For color clustering
matplotlib>=3.8.0  # For color palette visualization 